image-layout: add an initial image layout spec
In order to build tooling that converts the set of OCI Image objects
into an OCI Runtime object we need an Image Layout that test tools can
be built around.

Longer term this Image Layout could be used to define a "single object
image" that is a tar/zip/etc that could be posted over https/ftp/etc.

The layout comes from several different discussions:

https://github.com/opencontainers/image-spec/issues/23
https://github.com/opencontainers/image-spec/issues/92

Signed-off-by: Brandon Philips <brandon.philips@coreos.com>