manifest: List media types supported by descriptor references
Require implementations to support these media types (although an
implementation which also supports additional media types is still
compliant).  Suggest portable manifests and manifest lists stick to
these types so the can rely on implementation support.

A manifest(-list) can safely use a type from outside the list if the
author expects the implementation to support the extension type
(e.g. application/vnd.docker.image.rootfs.diff.tar.gzip).  But how
that out-of-band expectation is setup is beyond the scope of this
specification.

Signed-off-by: W. Trevor King <wking@tremily.us>