README: Link to specs-go and schema and declare Go-compat policy
The links help with discoverability, otherwise folks reading the
README might not notice that we provided these resources in addition
to the spec itself.

By declaring a Go-compat policy, folks who have Go troubles can tell
without testing whether the image-spec tooling *should* work for their
Go environment.  And if/when it does not, they can see whether
image-spec is interested in patches or not.  For example, if the
tooling breaks on Go 1.6, we don't care or want some awkward
workaround.  But if it breaks on Go 1.7 we do care and want a patch.

The Go-compat policy formalizes [1].  Previous maintainer comments
suggested some support for older Go releases [2], and I personally
think we want to give people more flexibility (not everyone can
upgrade to a new Go version on the day it comes out), but this commit
at least documents where we are now as a base for future discussion.

[1]: https://github.com/opencontainers/image-spec/pull/500#issuecomment-267678301
[2]: https://github.com/opencontainers/image-spec/pull/487#issuecomment-265876398

Signed-off-by: W. Trevor King <wking@tremily.us>