image-layout: Drop manifest ref example description
This ref -> descriptor example caption gained "blobs the manifest
references" wording with 2f24791 (image-layout.md: explain blobs can
be anything, 2016-06-14, #136), but I don't see a direct relationship
between descriptors and those deeper-ancestor blobs.  I think we
either want to revert the changes 2f24791 made to this line or drop
the line.  In this commit I drop the line, because the other points
covered by the sentence are already covered in other descriptor docs.
And it's pretty clear just from the example command and output that
this is showing the content of a ref pointing at a manifest list.

Signed-off-by: W. Trevor King <wking@tremily.us>